[FEAT] Dino ErrorCode & Test 추가

+++ src/main/java/kr/sesac/aoao/server/dino/controller/dto/GetUserDinoResponse.java
@@ -3,6 +3,7 @@ package kr.sesac.aoao.server.dino.controller.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+
 @Getter
 public class GetUserDinoResponse {
 
@@ -10,11 +11,11 @@ public class GetUserDinoResponse {
 	private final String name;
 	private final String color;
 	private final Integer exp;
-	private final Integer Lv;
+	private final Long Lv;
 	private final Integer point;
 
 	@Builder
-	public GetUserDinoResponse(Long userId, String name, String color, Integer exp, Integer lv, Integer point) {
+	public GetUserDinoResponse(Long userId, String name, String color, Integer exp, Long lv, Integer point) {
 		this.userId = userId;
 		this.name = name;
 		this.color = color;
